Move BranchesToMark into config package
